Update sources of metro hash.

+++ benchmarks/metrohash64.cpp
@@ -1,31 +1,23 @@
 // metrohash64.cpp
 //
-// The MIT License (MIT)
+// Copyright 2015-2018 J. Andrew Rogers
 //
-// Copyright (c) 2015 J. Andrew Rogers
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
 //
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
+//     http://www.apache.org/licenses/LICENSE-2.0
 //
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
-#include <string.h>
 #include "platform.h"
 #include "metrohash64.h"
+
+#include <cstring>
 
 const char * MetroHash64::test_string = "012345678901234567890123456789012345678901234567890123456789012";
 
@@ -67,7 +59,7 @@ void MetroHash64::Update(const uint8_t * const buffer, const uint64_t length)
         if (fill > length)
             fill = length;
 
-        memcpy(input.b + (bytes % 32), ptr, fill);
+        memcpy(input.b + (bytes % 32), ptr, static_cast<size_t>(fill));
         ptr   += fill;
         bytes += fill;
         
@@ -82,7 +74,7 @@ void MetroHash64::Update(const uint8_t * const buffer, const uint64_t length)
     }
     
     // bulk update
-    bytes += (end - ptr);
+    bytes += static_cast<uint64_t>(end - ptr);
     while (ptr <= (end - 32))
     {
         // process directly from the source, bypassing the input buffer
@@ -94,7 +86,7 @@ void MetroHash64::Update(const uint8_t * const buffer, const uint64_t length)
     
     // store remaining bytes in input buffer
     if (ptr < end)
-        memcpy(input.b, ptr, end - ptr);
+        memcpy(input.b, ptr, static_cast<size_t>(end - ptr));
 }
